tools/cascade-research/pyproject.toml
[project]
name = "cascade-research"
version = "0.1.0"
description = "Semantic search CLI for Cascade research files"
requires-python = ">=3.9"
dependencies = [
    "chromadb>=1.0.0",
    "langchain-text-splitters>=0.2.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
cascade-research = "cascade_research.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
